backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.127.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "rich>=14.2.0",
    "sqlmodel>=0.0.29",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "python-multipart>=0.0.18",
    "uvicorn==0.24.0",
    "python-dotenv==1.0.0",
    "openai-agents==0.6.2",
]
